pull some code into separate modules, remove jquery, and resize game on window resize
