+++ package.json
@@ -21,7 +21,5 @@
     "browserify": "^12.0.1",
     "nodemon": "^1.8.1"
   },
-  "dependencies": {
-    "jquery": "^2.1.4"
-  }
+  "dependencies": {}
 }
